pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doctr-synth-generator"
description = "A synthetic data generator for training OCR models"
authors = [{name = "Felix Dittrich", email = "felixdittrich92@gmail.com"}]
maintainers = [
    {name = "Felix Dittrich"},
]
readme = "README.md"
requires-python = ">=3.10.0,<4"
license = {file = "LICENSE"}
keywords=["OCR", "deep learning", "computer vision", "text recognition", "synthetic data", "data augmentation", "image processing"]
classifiers=[
        "Development Status :: 4 - Beta",
        "Intended Audience :: Developers",
        "Intended Audience :: Education",
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: Apache Software License",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version"]
dependencies = [
    "fonttools>=4.50.0,<5.0.0",
    "numpy>=2.0.0,<3.0.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
testing = [
    "pytest>=5.3.2",
    "coverage[toml]>=4.5.4",
]
quality = [
    "ruff>=0.1.5",
    "mypy>=0.812",
    "pre-commit>=2.17.0",
]
dev = [
    # core
    "fonttools>=4.50.0,<5.0.0",
    "numpy>=2.0.0,<3.0.0",
    "Pillow>=11.0.0",
    # testing
    "pytest>=5.3.2",
    "coverage[toml]>=4.5.4",
    # quality
    "ruff>=0.1.5",
    "mypy>=0.812",
    "pre-commit>=2.17.0",
]

[project.urls]
repository = "https://github.com/felixdittrich92/docTR-Synth-Generator"
tracker = "https://github.com/felixdittrich92/docTR-Synth-Generator/issues"
changelog = "https://github.com/felixdittrich92/docTR-Synth-Generator/releases"

[tool.setuptools]
zip-safe = true

[tool.setuptools.packages.find]
exclude = ["tests*"]

[tool.setuptools.package-data]
generator = ["py.typed"]

[tool.mypy]
files = "generator/"
show_error_codes = true
pretty = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true
implicit_reexport = false

[[tool.mypy.overrides]]
module = [
	"fontTools.*",
    "numpy.*",
    "PIL.*",
]
ignore_missing_imports = true

[tool.ruff]
exclude = [".git", "venv*", "build", "**/__init__.py"]
line-length = 120
target-version = "py310"
preview=true

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/
    "E", "W", "F", "I", "N", "Q", "C4", "T10", "LOG",
    "D101", "D103", "D201","D202","D207","D208","D214","D215","D300","D301","D417", "D419", "D207"  # pydocstyle
]
ignore = ["E402", "E203", "F403", "E731", "N812", "N817", "C408", "LOG015"]

[tool.ruff.lint.isort]
known-first-party = ["generator"]
known-third-party = ["fontTools", "numpy", "PIL"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["D"]
".github/**.py" = ["D"]


[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.coverage.run]
source = ["generator"]
